Took the checking of invariants 2 and 3 out of the add operation and moved them into a the insertleaf function in order to make the transfers of blocks1 into blocks2 every time it is needed instead of only when a block1 is full.
Make it transfer the whole pair of blocks1 from one block2 to another, instead of just one block2 and get problems with the mate being in another block2.

Changed the initialisation of OldNode and NewNode to that oldnode is null and newnode is the node that this block1 is actually in.